Changed type of DF2QueryResult.intentDetectionConfidence to float.
This type change was made in order to reflect the official Google DF2 APIs, since that field can contain continuous values in the [0,1] interval, and its parsing process can cause number format exceptions.

+++ Assets/Scripts/DialogflowV2Agents/DF2QueryResult.cs
@@ -31,7 +31,7 @@ namespace Syrus.Plugins.DFV2Client
 		public Dictionary<string, object> intent { get; set; }
 
 		[JsonProperty]
-		public int intentDetectionConfidence { get; set; }
+		public float intentDetectionConfidence { get; set; }
 
 		[JsonProperty]
 		public string languageCode { get; set; }
